+++ src/plugins/dexie/rx-storage-instance-dexie.ts
@@ -373,6 +373,7 @@ export class RxStorageInstanceDexie<RxDocType> implements RxStorageInstance<
                                     doc: null
                                 };
                             } else if (documentInDb._deleted && docData._deleted) {
+                                bulkPutDocs.push(docData);
                                 change = null;
                             }
                             if (change) {